requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
filelock==3.25.2
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
python-whois==0.9.6
requests==2.32.5
requests-file==3.0.1
six==1.17.0
tldextract==5.3.1
urllib3==2.6.3
validators==0.35.0
Werkzeug==3.1.6
